updating the code

+++ src/main/java/com/example/wmsnew/warehouse/repository/LocationRepository.java
@@ -6,4 +6,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
 public interface LocationRepository
-    extends JpaRepository<Location, Integer>, JpaSpecificationExecutor<Location> {}
+    extends JpaRepository<Location, Long>, JpaSpecificationExecutor<Location> {}
